backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sachin <sachin2262716@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7"
Django = "^3.1.6"
djangorestframework = "^3.12.2"
PyMySQL = "^1.0.2"
Pillow = "^8.1.0"
Faker = "^5.8.0"
django-cors-headers = "^3.7.0"
cryptography = "^3.4.6"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
